Refactor database

See merge request 2dv612/student/team-2/project!51

+++ database/sql/user.sql
@@ -59,7 +59,7 @@ CREATE TABLE product (
   producer_org_no VARCHAR(10) NOT NULL,
   name VARCHAR(100) NOT NULL,
   description VARCHAR(500),
-  price DOUBLE NOT NULL,
+  price INT NOT NULL,
   unit VARCHAR(20) NOT NULL,
   in_stock INT NOT NULL,
   PRIMARY KEY (id),
